makefile rework

+++ libft/srcs/ft_split.c
@@ -31,7 +31,7 @@ static int	ft_cntstr(char const *s, char c)
 	return (n);
 }
 
-static void	ft_free(char **str, size_t i)
+static void	ft_free(char **str, int i)
 {
 	while (i >= 0)
 	{
@@ -41,10 +41,10 @@ static void	ft_free(char **str, size_t i)
 	free(str);
 }
 
-static char	**ft_allocate(char **new, const char *s, char c, size_t i)
+static char	**ft_allocate(char **new, const char *s, char c, int i)
 {
-	size_t	k;
-	size_t	start;
+	int	k;
+	int	start;
 
 	k = 0;
 	while (s[i])
@@ -72,7 +72,7 @@ static char	**ft_allocate(char **new, const char *s, char c, size_t i)
 char	**ft_split(char const *s, char c)
 {
 	char	**new;
-	size_t	i;
+	int		i;
 
 	if (s == NULL)
 		return (NULL);
